Fix wind arrow
